delete Enrollment implementieren. deleteByRatingID im ApplicationMapper implementieren.

+++ src/de/worketplace/team06/server/WorketplaceAdministrationImpl.java
@@ -299,23 +299,27 @@ public class WorketplaceAdministrationImpl extends RemoteServiceServlet implemen
 				//Für jede Ausschreibung wird die dazugehörige Bewertung ausgelesen
 				Rating rating = ratingMapper.findRatingByApplicationID(a.getID());
 					
+				if (rating != null){
 				//Es wird überprüft ob die Bewertung auch mit einer Enrollment Instanz verbunden ist
 				//Wenn ja wird der Rating Mapper nicht gelöscht, Wenn nein wird er gelöscht. 
 					if (this.enrollMapper.findByRatingID(rating.getID()) == null){
 						this.ratingMapper.delete(rating);
-						this.appMapper.delete(a);
 					}
 					
-					else{
-						this.appMapper.delete(a);
-					}
 					
+				}	
+				else {
+					//TO Do: Fehlermeldung
+				}
+				this.appMapper.delete(a);
 											
 			}
 		}
 		
 		PartnerProfile pp = getPartnerProfileFor(call);
 		if (pp != null){
+			
+			this.propertyMapper.deleteByPartnerProfileID(pp.getID());
 			this.partnerMapper.delete(pp);
 		} else {
 			//TODO: Fehlermeldung
@@ -405,6 +409,12 @@ public class WorketplaceAdministrationImpl extends RemoteServiceServlet implemen
 	 */
 	@Override
 	public void deleteEnrollment(Enrollment enrollment) throws IllegalArgumentException {
+		Rating r = this.ratingMapper.findById(enrollment.getRatingID());
+		
+		if (r != null && this.appMapper.findByRatingID(r.getID()) == null){
+			this.ratingMapper.delete(r);
+		}
+		
 		this.enrollMapper.delete(enrollment);
 		
 	}
